move signature gen to firebase func
