+++ frontend/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@google-cloud/secret-manager": "^4.1.4",
     "@headlessui/react": "^1.7.4",
     "@heroicons/react": "^2.0.13",
     "@moralisweb3/client-firebase-auth-utils": "^2.7.1",
