+++ functions/index.js
@@ -2,6 +2,10 @@ const functions = require("firebase-functions");
 const { Client, resources, Webhook } = require("coinbase-commerce-node");
 const { Charge } = resources;
 const admin = require("firebase-admin");
+const { ethers } = require("ethers");
+const { defaultAbiCoder } = ethers.utils;
+
+// RUN LOCALLY: firebase emulators:start --only functions
 
 // setup cors
 const cors = require("cors")({ origin: "*" });
@@ -14,6 +18,7 @@ const firestore = admin.firestore();
 require("dotenv").config();
 const cbApiKey = process.env.COINBASE_COMMERCE_API_KEY;
 const cbWebhookSecret = process.env.COINBASE_COMMERCE_WEBHOOK_SECRET;
+const signerPrivateKey = process.env.COMMUNITY_SIGNER_PRIVATE_KEY;
 
 Client.init(cbApiKey);
 
@@ -157,3 +162,40 @@ exports.webhookHandler = functions.https.onRequest(
     }
   }
 );
+
+exports.genSignature = functions.https.onRequest((request, response) => {
+  // to mint community art we need a securely generated signature (to prevent bot abuse)
+  // to instantiate the signer we need a private key, which is stored on google's secret manager
+  // this can only be securely accesses server-side, so it needs to run in a firebase function
+  cors(request, response, async () => {
+    console.log("generating signature", request.body);
+    try {
+      const { contractAddress, allowlistAddress, balance } = JSON.parse(
+        request.body
+      );
+      if (!contractAddress || !allowlistAddress) {
+        throw new Error("Missing contractAddress or allowlistAddress");
+      }
+      // get the signer (this address should be added to the contract as a signer)
+      const signer = new ethers.Wallet(signerPrivateKey);
+      console.log("signer", signer.address);
+      const payload = defaultAbiCoder.encode(
+        ["address", "address", "uint256"],
+        [contractAddress, allowlistAddress, balance]
+      );
+      console.log("signing");
+      const signature = await signer.signMessage(
+        ethers.utils.arrayify(payload)
+      );
+      console.log("signature", signature);
+      response.status(200).json({ signature });
+    } catch (error) {
+      console.log * ("error", error);
+      try {
+        response.status(500).json({ error: error.message });
+      } catch {
+        response.status(500).json({ unknown_error: error });
+      }
+    }
+  });
+});
