+++ functions/package.json
@@ -16,6 +16,7 @@
     "coinbase-commerce-node": "^1.0.4",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
+    "ethers": "^5.7.2",
     "firebase-admin": "^10.3.0",
     "firebase-functions": "^4.0.2"
   },
